code/requirements.txt
# ── Core LLM Framework ────────────────────────────────────
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0

# ── Web UI ────────────────────────────────────────────────
streamlit>=1.40.0

# ── Environment & Config ─────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.0.0

# ── RAG: Codebase Context (Feature 1) ────────────────────
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0

# ── Git Integration (Feature 4) ──────────────────────────
gitpython>=3.1.0
